requirements.txt
mflux>=0.1.0
kokoro-onnx>=0.1.8
faster-whisper>=1.0.0
ollama>=0.1.6
pyyaml>=6.0.1
soundfile>=0.12.1
numpy>=1.26.4
torch>=2.2.0
diffusers>=0.27.2
transformers>=4.40.0
accelerate>=0.29.3
librosa>=0.10.1
Pillow>=10.3.0
